cloudsql_backend/requirements.txt
fastapi
uvicorn[standard]
google-cloud-secret-manager
firebase-admin
cloud-sql-python-connector[pg8000]
psycopg2-binary
google-auth
google-cloud-storage